High Contrast Fixes (#10719)
This primarily updates areas in pxt where we were checking the old true/false high contrast setting (which is no longer used) instead of the using the new theme manager. It was just an oversight in the theming work. I've created a little static helper function in the theme manager to save some repetitive lines of code.

Separately, it fixes a few places where we were seeing superfluous yellow outlines in hc mode. We have some generic highlighting css logic which, in general, is convenient. But it does seem to be causing some issues too, highlighting things we don't want highlighted. I've special-cased those out for now, but if it becomes a more common occurrence, we may need to rework it more completely.