Do not show debug suggestion when a project can't compile (#10721)
For text errors, we were able to determine if an error was a compile or runtime error using the presence of a stack trace, but for block errors, there are some that do not prevent running the program but also do not have a stack, so the same method was not as reliable (hence, we just always showed the button).

However, looking more closely, I believe we can determine if it's a run-blocking error based on whether we detect the error in the typescript compile diagnostics or not. With that in mind, I've added a specific preventsRunning flag to errors, which we can now set and use for determining if we should show the debugger suggestion. When we detect a typescript compile error in blocks, we set this to true. If it's a runtime or blocks compilation error, we set it to false. I've switched the monaco errors to use this as well, though the actual behavior is unchanged.