Stop simulator from closing dropdowns when restarted from melodyeditor (#10735)
The melody editor suspends the simulator while it is open, to prevent audio issues. However, restarting contains a call to close any blockly dropdowns, which is a problem if you closed the melody editor by clicking on a dropdown.

This plumbs through the background flag to enable the melody editor to restart the simulator without closing any unnecessary dropdowns.