Fix not showing Keyboard Controls in help menu
It was missing on first load because there isn't an editor at that point so
isBlocksEditor returns false. Nothing causes the menu to re-render when an
editor has been assigned. The settings component had a working inBlocks prop
so we've switched to use that approach. I think that only really works because
the parent component of the menus is not pure. So far as I can tell, there's no
React state for the current editor.

Also simplify the conditions which had become over complicated since the
decision was made to require the blocks editor to be active, not just the
project to use blocks.