Reload code eval iframe even when URL is the same (#10782) (#10784)
This change provides a way for the user to trigger a reload of the project when it occurs.

Previously, if the shareId didn't change, the project wouldn't reload, even if you clicked enter in the input field. With this change, we force the reload even if shareId remains the same. This was a little weird to do, since we rely on the shareId state update to trigger the normal iframe reload, but by adding a simple counter to increment in addition to the URL, we can force that code to run even if shareId is the same.