Always unregister and reregister copy paste handlers in initCopyPaste
These handlers can be unregistered when we clean up keyboard nav (and not reinstated), so when we clean up & re-init keyboard nav, we'd end up in a state where oldCopy is set but the shortcut is no longer registered, so initCopyPaste is skipped even though we actually need it to run. And that'd cause failures down the line in the re-initialization for keyboard nav.