Deal with more conflicting shortcut registrations.
Was seeing issues going back and forth between js and blocks. Initially, it was failing because the keyboard_nav shortcuts were unregistered in initCopyPaste and not re-registered, but seemed to still be expected by some parts of the code. After removing the unregister from there, however, it would fail when the keyboard_nav shortcuts *did* exist and initCopyPaste tried to add a conflicting shortcut.

To address this, go back to making initCopyPaste no-op if oldCopy exists (so we don't get conflicting registrations), but add a flag to force re-initialization, which we can use after keyboard nav has been cleaned up (so we can re-create the old copy/paste handlers after keyboardNav cleanup removes them).