Fix Keyboard Nav with FlyoutOnly Mode (#10871)
* Check for flyoutOnly when getting toolbox bounds in area menu overlay

* Fix focusing on the toolbox in flyoutOnly mode

* Another fixed check for cachingflyout

* Always unregister and reregister copy paste handlers in initCopyPaste

These handlers can be unregistered when we clean up keyboard nav (and not reinstated), so when we clean up & re-init keyboard nav, we'd end up in a state where oldCopy is set but the shortcut is no longer registered, so initCopyPaste is skipped even though we actually need it to run. And that'd cause failures down the line in the re-initialization for keyboard nav.

* Move registerKeyboardNavigationStyles into loadBlocklyAsync. As far as I can tell, this only runs once (unlike prepareBlockly), so it removed the need for the separate registeredKeyboardNavigationStyles flag.

* Use existing function to move focus to flyout

* Add keyboard navigation cleanup step to refreshBlockly

* Patch to ensure we cannot register duplicate key bindings

* Remove extra import

* Includes instead of indexOf

* Update hasCategories check to look specifically for false. This can be undefined when the toolbox has categories (for example, in microbit).

* A few more unregisters needed during initCopyPaste

* Deal with more conflicting shortcut registrations.

Was seeing issues going back and forth between js and blocks. Initially, it was failing because the keyboard_nav shortcuts were unregistered in initCopyPaste and not re-registered, but seemed to still be expected by some parts of the code. After removing the unregister from there, however, it would fail when the keyboard_nav shortcuts *did* exist and initCopyPaste tried to add a conflicting shortcut.

To address this, go back to making initCopyPaste no-op if oldCopy exists (so we don't get conflicting registrations), but add a flag to force re-initialization, which we can use after keyboard nav has been cleaned up (so we can re-create the old copy/paste handlers after keyboardNav cleanup removes them).