Fix keyboard nav to Basic category (#10872)
Since the basic category (top blocks) wasn't special cased and not treated as a normal item in the toolbox, we'd skip over it and go straight to search when navigating up. To fix this, we can just include the topblocks category alongside all the others, so it all behaves consistently.